# crates/capsem-service/src/router_runtime.rs
use super::*;

pub(crate) mod exposures;
pub(crate) mod restart;
mod streams;

pub(super) fn build_service_router(state: Arc<ServiceState>) -> Router {
    Router::new()
        .route("/status", get(handle_service_status))
        .route("/restart", post(restart::handle_restart))
        .route("/update/status", get(handle_update_status))
        .route("/system/status", get(handle_system_status))
        .route("/update/check", post(handle_update_check))
        .route("/update/apply", post(handle_update_apply))
        .route(
            "/version",
            get(|| async { Json(serde_json::json!({ "version": env!("CARGO_PKG_VERSION") })) }),
        )
        .route(
            "/networks",
            get(network_routes::handle_networks_list).post(network_routes::handle_network_create),
        )
        .route(
            "/networks/{id}",
            get(network_routes::handle_network_inspect).delete(network_routes::handle_network_delete),
        )
        .route("/networks/{id}/logs", get(network_routes::handle_network_logs))
        .route(
            "/networks/private/resolve",
            post(private_routes::handle_private_resolve),
        )
        .route(
            "/networks/{id}/members/{vm_id}",
            put(network_routes::handle_network_attach).delete(network_routes::handle_network_detach),
        )
        .route("/images", get(container_setup::images::handle_list_images))
        .route("/images/pull", post(container_setup::images::handle_pull_image))
        .route("/vms/create", post(handle_provision))
        .route("/vms/list", get(handle_list))
        .route("/vms/{id}/info", get(handle_info))
        .route("/vms/{id}/status", get(handle_vm_status))
        .route("/vms/{id}/container", get(container_setup::handle_container_status))
        .route("/vms/{id}/stream", get(streams::handle_stream))
        .route(
            "/vms/{id}/exposures",
            get(exposures::handle_list_exposures).post(exposures::handle_create_exposure),
        )
        .route(
            "/vms/{id}/exposures/{exposure_id}",
            delete(exposures::handle_delete_exposure),
        )
        .route(
            "/vms/{id}/exposures/{exposure_id}/preview-session",
            delete(exposures::handle_revoke_preview_sessions),
        )
        .route(
            "/internal/vms/{id}/exposures/{exposure_id}/preview-session",
            post(exposures::handle_create_preview_session),
        )
        .route(
            "/internal/vms/{id}/exposures/{exposure_id}/preview-bootstrap",
            post(exposures::handle_exchange_preview_bootstrap),
        )
        .route(
            "/internal/vms/{id}/exposures/{exposure_id}/preview-admission",
            post(exposures::handle_admit_preview_connection),
        )
        .route("/vms/{id}/logs", get(handle_logs))
        .route("/vms/{id}/exec", post(handle_exec))
        .route("/vms/{id}/stop", post(handle_stop))
        .route("/vms/{id}/pause", post(handle_suspend))
        .route("/vms/{id}/delete", delete(handle_delete))
        .route("/vms/{id}/preserve-failure", post(handle_preserve_failure))
        .route("/vms/{id}/start", post(handle_resume))
        .route("/vms/{id}/resume", post(handle_resume))
        .route("/vms/{id}/save", post(handle_persist))
        .route("/vms/{id}/save/status", get(handle_vm_save_status))
        .route("/vms/{id}/fork/status", get(handle_vm_fork_status))
        .route("/purge", post(handle_purge))
        .route("/run", post(handle_run))
        .route("/stats", get(handle_stats))
        .route("/vms/{id}/stats/summary", get(handle_stats_summary))
        .route("/vms/{id}/stats/detail", get(handle_stats_detail))
        .route("/service-logs", get(handle_service_logs))
        .route("/triage", get(handle_triage))
        .route("/panics", get(handle_panics))
        .route("/host-logs/{name}", get(handle_host_logs))
        .route("/vms/{id}/bodies/export.warc.gz", get(handle_bodies_warc_export))
        .route("/vms/{id}/bodies/{event_id}", get(handle_event_bodies))
        .route("/vms/{id}/timeline", get(handle_timeline))
        .route("/vms/{id}/security/latest", get(handle_security_latest))
        .route("/vms/{id}/security/status", get(handle_security_info))
        .route("/vms/{id}/detection/latest", get(handle_detection_latest))
        .route("/vms/{id}/detection/status", get(handle_security_info))
        .route("/vms/{id}/enforcement/latest", get(handle_security_latest))
        .route("/vms/{id}/enforcement/status", get(handle_security_info))
        .route("/security/latest", get(handle_service_security_latest))
        .route("/security/status", get(handle_service_security_status))
        .route("/enforcement/latest", get(handle_service_security_latest))
        .route("/enforcement/status", get(handle_service_security_status))
        .route("/detection/latest", get(handle_service_detection_latest))
        .route("/detection/status", get(handle_service_detection_status))
        .route("/assets/status", get(asset_routes::handle_asset_status))
        .route("/assets/ensure", post(asset_routes::handle_asset_ensure))
        .route("/plugins/list", get(plugin_routes::handle_plugins))
        .route(
            "/plugins/credential_broker/credentials/info",
            get(plugin_routes::handle_credential_broker_credentials_info),
        )
        .route(
            "/plugins/credential_broker/credentials/reload",
            post(plugin_routes::handle_credential_broker_credentials_reload),
        )
        .route("/plugins/{plugin_id}/info", get(plugin_routes::handle_plugin_info))
        .route("/plugins/{plugin_id}/edit", patch(plugin_routes::handle_plugin_update))
        .route("/vms/{id}/fork", post(handle_fork))
        .route("/settings/info", get(handle_get_settings))
        .route("/settings/edit", patch(handle_save_settings))
        .route("/corp/info", get(handle_corp_info))
        .route("/corp/edit", put(handle_corp_config))
        .route("/corp/validate", post(handle_corp_validate))
        .route("/corp/reload", post(handle_corp_reload))
        .route("/mcp/info", get(mcp_routes::handle_mcp_info))
        .route("/mcp/servers/list", get(mcp_routes::handle_mcp_servers))
        .route("/mcp/default/info", get(mcp_routes::handle_mcp_default_info))
        .route("/mcp/default/edit", patch(mcp_routes::handle_mcp_default_edit))
        .route(
            "/mcp/servers/{server_id}/tools/list",
            get(mcp_routes::handle_mcp_server_tools),
        )
        .route(
            "/mcp/servers/{server_id}/refresh",
            post(mcp_routes::handle_mcp_server_refresh),
        )
        .route(
            "/mcp/servers/{server_id}/tools/{tool_id}/edit",
            patch(mcp_routes::handle_mcp_tool_edit),
        )
        .route(
            "/mcp/servers/{server_id}/tools/{tool_id}/call",
            post(mcp_routes::handle_mcp_tool_call),
        )
        .route("/vms/{id}/history", get(handle_history))
        .route("/vms/{id}/history/processes", get(handle_history_processes))
        .route("/vms/{id}/history/counts", get(handle_history_counts))
        .route("/vms/{id}/history/transcript", get(handle_history_transcript))
        .route("/vms/{id}/files/list", get(handle_list_files))
        .route(
            "/vms/{id}/files/content",
            get(handle_download_file).post(handle_upload_file),
        )
        // Accept what the gateway forwards; axum's implicit 2 MiB default
        // refused file uploads the API documents.
        .layer(axum::extract::DefaultBodyLimit::max(capsem_api::MAX_REQUEST_BODY_BYTES))
        .layer(TraceLayer::new_for_http().on_request(()).on_response(()))
        .with_state(state)
}

pub(super) async fn handle_update_status(
    State(state): State<Arc<ServiceState>>,
) -> Result<Json<api::UpdateStatusResponse>, AppError> {
    state.off_worker(|state| update_status_response(&state)).await.map(Json)
}

pub(super) async fn handle_system_status(
    State(state): State<Arc<ServiceState>>,
) -> Result<Json<api::SystemStatusResponse>, AppError> {
    let (manifest, manifest_metadata, updates) = state
        .off_worker(|state| {
            let manifest = read_installed_status_document(&state.assets_dir.join("manifest.json"))?;
            let metadata = read_manifest_metadata_status_document(&state.assets_dir.join("manifest-metadata.json"))?;
            Ok::<_, AppError>((manifest, metadata, update_status_response(&state)))
        })
        .await??;
    capsem_assets::asset_manager::ManifestV2::from_json(&serde_json::to_string(&manifest).map_err(|error| {
        AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("serialize installed manifest for validation: {error}"),
        )
    })?)
    .map_err(|error| {
        AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("installed manifest is invalid: {error:#}"),
        )
    })?;
    let assets = state.off_worker(|state| asset_routes::asset_status(&state)).await??;
    Ok(Json(api::SystemStatusResponse {
        version: state.current_version.clone(),
        service: "running".to_string(),
        manifest,
        manifest_metadata,
        assets,
        corp: corp_info_value()?,
        updates,
    }))
}

pub(super) fn read_installed_status_document(path: &StdPath) -> Result<serde_json::Value, AppError> {
    let content = std::fs::read_to_string(path).map_err(|error| {
        AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("read installed status document {}: {error}", path.display()),
        )
    })?;
    serde_json::from_str(&content).map_err(|error| {
        AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("parse installed status document {}: {error}", path.display()),
        )
    })
}

pub(super) fn read_manifest_metadata_status_document(path: &StdPath) -> Result<serde_json::Value, AppError> {
    let value = read_installed_status_document(path)?;
    if value.get("schema").and_then(serde_json::Value::as_str) != Some("capsem.manifest_metadata.v1") {
        return Err(AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            "installed manifest metadata must use schema capsem.manifest_metadata.v1".to_string(),
        ));
    }
    Ok(value)
}

pub(super) async fn handle_update_check(
    State(state): State<Arc<ServiceState>>,
    Json(request): Json<api::UpdateCheckRequest>,
) -> Result<Json<api::UpdateActionResponse>, AppError> {
    let plan = update_command_plan(UpdateCommandKind::Check);
    if request.dry_run {
        return Ok(Json(planned_update_response(plan)));
    }
    execute_update_command(&state, plan).await.map(Json)
}

pub(super) async fn handle_update_apply(
    State(state): State<Arc<ServiceState>>,
    Json(request): Json<api::UpdateApplyRequest>,
) -> Result<Json<api::UpdateActionResponse>, AppError> {
    let plan = update_command_plan(UpdateCommandKind::Apply);
    if request.dry_run {
        return Ok(Json(planned_update_response(plan)));
    }
    if !request.confirmed {
        return Err(AppError(
            StatusCode::BAD_REQUEST,
            "update apply requires confirmed=true or dry_run=true".to_string(),
        ));
    }
    execute_update_apply(&state, plan).await.map(Json)
}

pub(super) fn planned_update_response(plan: api::UpdateCommandPlan) -> api::UpdateActionResponse {
    api::UpdateActionResponse {
        status: api::UpdateActionStatus::Planned,
        command: plan,
        exit_code: None,
        stdout: None,
        stderr: None,
    }
}

pub(super) async fn execute_update_command(
    state: &ServiceState,
    plan: api::UpdateCommandPlan,
) -> Result<api::UpdateActionResponse, AppError> {
    let _update_guard = state.update_lock.lock().await;
    execute_update_command_unlocked(plan).await
}

pub(super) async fn execute_update_apply(
    state: &ServiceState,
    plan: api::UpdateCommandPlan,
) -> Result<api::UpdateActionResponse, AppError> {
    let _update_guard = state.update_lock.lock().await;
    let response = execute_update_command_unlocked(plan).await?;
    if response.status == api::UpdateActionStatus::Succeeded {
        reload_activated_update_runtime(state)?;
    }
    Ok(response)
}

pub(super) async fn execute_update_command_unlocked(
    plan: api::UpdateCommandPlan,
) -> Result<api::UpdateActionResponse, AppError> {
    let output = Command::new(&plan.program)
        .args(&plan.args)
        .output()
        .await
        .map_err(|error| {
            AppError(
                StatusCode::INTERNAL_SERVER_ERROR,
                format!("failed to start update command: {error}"),
            )
        })?;
    let status = if output.status.success() {
        api::UpdateActionStatus::Succeeded
    } else {
        api::UpdateActionStatus::Failed
    };
    Ok(api::UpdateActionResponse {
        status,
        command: plan,
        exit_code: output.status.code(),
        stdout: Some(String::from_utf8_lossy(&output.stdout).to_string()),
        stderr: Some(String::from_utf8_lossy(&output.stderr).to_string()),
    })
}

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
pub(super) enum UpdateRuntimeDisposition {
    Reloaded,
    RestartRequested,
}

#[derive(Debug, Clone, PartialEq, Eq)]
pub(super) enum AutomaticUpdateOutcome {
    Disabled,
    Busy,
    Succeeded(UpdateRuntimeDisposition),
    Failed(String),
}

pub(super) fn automatic_updates_enabled() -> bool {
    let (user, corp) = capsem_core::net::policy_config::load_settings_and_corp_files();
    let resolved = capsem_core::net::policy_config::resolve_settings(&user, &corp);
    automatic_updates_enabled_from_resolved(&resolved)
}

pub(super) fn automatic_updates_enabled_from_resolved(
    settings: &[capsem_core::net::policy_config::ResolvedSetting],
) -> bool {
    settings
        .iter()
        .find(|setting| setting.id == "app.auto_update")
        .and_then(|setting| setting.effective_value.as_bool())
        .unwrap_or(true)
}

pub(super) fn automatic_update_failure_backoff(consecutive_failures: u32) -> std::time::Duration {
    let exponent = consecutive_failures.saturating_sub(1).min(8);
    let seconds = AUTOMATIC_UPDATE_POLL_SECS
        .saturating_mul(1_u64 << exponent)
        .min(AUTOMATIC_UPDATE_MAX_BACKOFF_SECS);
    std::time::Duration::from_secs(seconds)
}

pub(super) fn automatic_update_delay_from_value(
    value: Option<&std::ffi::OsStr>,
    default_seconds: u64,
) -> std::time::Duration {
    let seconds = value
        .and_then(std::ffi::OsStr::to_str)
        .and_then(|value| value.parse::<u64>().ok())
        .filter(|seconds| *seconds > 0)
        .unwrap_or(default_seconds);
    std::time::Duration::from_secs(seconds)
}

pub(super) fn automatic_update_delay(environment_name: &str, default_seconds: u64) -> std::time::Duration {
    let value = std::env::var_os(environment_name);
    automatic_update_delay_from_value(value.as_deref(), default_seconds)
}

pub(super) async fn run_automatic_update_once(state: &ServiceState) -> AutomaticUpdateOutcome {
    if !automatic_updates_enabled() {
        return AutomaticUpdateOutcome::Disabled;
    }
    let Ok(_update_guard) = state.update_lock.try_lock() else {
        return AutomaticUpdateOutcome::Busy;
    };
    let plan = update_command_plan(UpdateCommandKind::Apply);
    let response = match execute_update_command_unlocked(plan).await {
        Ok(response) => response,
        Err(error) => return AutomaticUpdateOutcome::Failed(error.body.error),
    };
    if response.status != api::UpdateActionStatus::Succeeded {
        let detail = response
            .stderr
            .as_deref()
            .map(str::trim)
            .filter(|detail| !detail.is_empty())
            .unwrap_or("update command returned a non-zero exit status");
        return AutomaticUpdateOutcome::Failed(format!("exit {:?}: {detail}", response.exit_code));
    }
    match reload_activated_update_runtime(state) {
        Ok(disposition) => AutomaticUpdateOutcome::Succeeded(disposition),
        Err(error) => AutomaticUpdateOutcome::Failed(error.body.error),
    }
}

pub(super) async fn run_automatic_update_loop(state: Arc<ServiceState>) {
    let mut delay = automatic_update_delay(AUTOMATIC_UPDATE_INITIAL_DELAY_ENV, AUTOMATIC_UPDATE_INITIAL_DELAY_SECS);
    let poll_delay = automatic_update_delay(AUTOMATIC_UPDATE_POLL_ENV, AUTOMATIC_UPDATE_POLL_SECS);
    let mut consecutive_failures = 0_u32;
    loop {
        tokio::time::sleep(delay).await;
        match run_automatic_update_once(&state).await {
            AutomaticUpdateOutcome::Disabled => {
                consecutive_failures = 0;
                delay = poll_delay;
                info!("automatic release polling is disabled by app.auto_update");
            }
            AutomaticUpdateOutcome::Busy => {
                delay = std::time::Duration::from_secs(AUTOMATIC_UPDATE_BUSY_RETRY_SECS);
                info!("automatic release polling skipped because an explicit update owns the lock");
            }
            AutomaticUpdateOutcome::Succeeded(disposition) => {
                consecutive_failures = 0;
                delay = poll_delay;
                info!(
                    ?disposition,
                    next_poll_secs = poll_delay.as_secs(),
                    "automatic release update completed"
                );
            }
            AutomaticUpdateOutcome::Failed(error) => {
                consecutive_failures = consecutive_failures.saturating_add(1);
                delay = automatic_update_failure_backoff(consecutive_failures);
                warn!(
                    error = %error,
                    consecutive_failures,
                    retry_secs = delay.as_secs(),
                    "automatic release update failed"
                );
            }
        }
    }
}

pub(super) fn should_start_automatic_update_loop(parent_pid: Option<u32>) -> bool {
    // `--parent-pid` is the explicit bounded-test harness rail. Production
    // services never receive it, while long VM suites must not gain an
    // unrelated network/package mutation one minute into a test process.
    parent_pid.is_none()
}

pub(super) fn reload_activated_update_runtime(state: &ServiceState) -> Result<UpdateRuntimeDisposition, AppError> {
    let path = state.assets_dir.join("manifest.json");
    let content = std::fs::read_to_string(&path).map_err(|error| {
        AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("read activated update manifest {}: {error}", path.display()),
        )
    })?;
    let manifest = capsem_assets::asset_manager::ManifestV2::from_json(&content).map_err(|error| {
        AppError(
            StatusCode::INTERNAL_SERVER_ERROR,
            format!("validate activated update manifest {}: {error:#}", path.display()),
        )
    })?;
    let selected_binary = manifest.binaries.current.clone();
    state
        .manifest
        .write()
        .map_err(|error| {
            AppError(
                StatusCode::INTERNAL_SERVER_ERROR,
                format!("installed manifest lock poisoned: {error}"),
            )
        })?
        .replace(Arc::new(manifest));
    // Resume state is derived from the installed manifest (revoked pins).
    state.persistent_resume_state_cache.lock().unwrap().clear();

    if selected_binary != state.current_version {
        info!(
            running = %state.current_version,
            selected = %selected_binary,
            "activated binary differs from the running service; requesting managed restart"
        );
        state.update_restart.notify_one();
        Ok(UpdateRuntimeDisposition::RestartRequested)
    } else {
        info!(
            selected = %selected_binary,
            "reloaded the activated manifest in the running service"
        );
        Ok(UpdateRuntimeDisposition::Reloaded)
    }
}

pub(super) async fn handle_service_status(
    State(state): State<Arc<ServiceState>>,
) -> Result<Json<serde_json::Value>, AppError> {
    let credential_store = capsem_core::credential_broker::credential_store_status();
    let ready = credential_store.ready;
    Ok(Json(serde_json::json!({
        "service": "capsem-service",
        "version": state.current_version,
        "ready": ready,
        "components": {
            "credential_store": {
                "ready": credential_store.ready,
                "status": credential_store.status,
                "last_error": credential_store.last_error,
            },
        },
    })))
}