# sdk/specification/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capsem Gateway",
    "version": "3.0.0"
  },
  "paths": {
    "/assets/status": {
      "get": {
        "operationId": "getAssetStatus",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStatus"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/host-logs/{name}": {
      "get": {
        "description": "Send Accept: application/json for the typed response. Without it the route returns plain text.",
        "operationId": "getHypervisorLogs",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/HostLogSource"
            }
          },
          {
            "name": "grep",
            "in": "query",
            "description": "Keep lines containing this literal substring, before applying tail.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tail",
            "in": "query",
            "description": "Return the last matching lines. Zero returns no lines.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "max_bytes",
            "in": "query",
            "description": "Read at most this many bytes per stream, capped at 5 MiB.\nDefaults to 100 KiB for host logs and 5 MiB for VM logs.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded, filtered log tail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostLogsResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/images": {
      "get": {
        "operationId": "listImages",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "description": "Read the catalog from its registry now instead of from the copy the\nservice keeps for up to 30 minutes.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageListResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/images/pull": {
      "post": {
        "operationId": "pullImage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImagePullRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImagePullResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/default/info": {
      "get": {
        "operationId": "getMcpDefault",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpDefaultPermissionResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/info": {
      "get": {
        "operationId": "getMcpInfo",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpInfoResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/servers/list": {
      "get": {
        "operationId": "listMcpServers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServersListResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/servers/{server_id}/refresh": {
      "post": {
        "operationId": "refreshMcpServer",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpRefreshResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/servers/{server_id}/tools/list": {
      "get": {
        "operationId": "listMcpTools",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpToolsListResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/servers/{server_id}/tools/{tool_id}/call": {
      "post": {
        "operationId": "callMcpTool",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Value"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Value"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/purge": {
      "post": {
        "operationId": "purgeVms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeResponse"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/status": {
      "get": {
        "operationId": "getHypervisorInfo",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HypervisorInfo"
                }
              }
            }
          },
          "default": {
            "description": "HTTP error; inspect status and error body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          },
          "timeout_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "vm_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "HostLogSource": {
        "type": "string",
        "description": "Symbolic host log streams. Paths supplied by clients are never accepted.",
        "enum": [
          "service",
          "mcp",
          "gateway",
          "tray",
          "app"
        ]
      },
      "HostLogsResponse": {
        "type": "object",
        "required": [
          "source",
          "text"
        ],
        "properties": {
          "source": {
            "$ref": "#/components/schemas/HostLogSource"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "McpServersListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/McpServerInfoResponse"
        }
      },
...
      "McpToolsListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/McpToolInfoResponse"
        }
      },
...
      "Value": {},
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}